+++ nix/tools/withTools.nix
@@ -25,7 +25,6 @@ let
             "ARG_USE_ENV([PGUSER], [postgrest_test_authenticator], [Authenticator PG role])"
             "ARG_USE_ENV([PGDATABASE], [postgres], [PG database name])"
             "ARG_USE_ENV([PGRST_DB_SCHEMAS], [test], [Schema to expose])"
-            "ARG_USE_ENV([PGRST_DB_ANON_ROLE], [postgrest_test_anonymous], [Anonymous PG role])"
           ];
         positionalCompletion = "_command";
         inRootDir = true;
@@ -54,7 +53,6 @@ let
         export PGUSER
         export PGDATABASE
         export PGRST_DB_SCHEMAS
-        export PGRST_DB_ANON_ROLE
 
         log "Initializing database cluster..."
         # We try to make the database cluster as independent as possible from the host